Feat : Loader 기능 추가
- 환율 정보 로딩중임을 사용자가 알 수 있도록 Loading animation UI를 추가
