+++ todo.md
@@ -3,7 +3,7 @@
 <hr>
 
 - [✅] 유로화를 입력할 때마다 환율을 반영하여 원화로 표시
-- [] 환율 정보 로딩중임을 사용자가 알 수 있도록 UI를 추가
+- [✅] 환율 정보 로딩중임을 사용자가 알 수 있도록 UI를 추가
 - [✅] 금액은 세자리 마다 콤마(,) 를 표시
 - [✅] 유로화 소수점 2자리까지 표시
 - [✅] 원화 소수점 표시하지 않음
